Hiram/requirements.txt
Django==5.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
python-dateutil==2.8.2 